% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-newbler.R
\name{add.gc}
\alias{add.gc}
\title{Extract GC via contig FASTA file and incorporate into len-cov data.frame}
\usage{
add.gc(contigSummData, fastaFile = "454LargeContigs.fna")
}
\arguments{
\item{contigSummData}{is a data.frame (returned from the function extract.contig.data)}

\item{fastaFile}{is a FASTA file containing the DNA sequences for the contigs in the assembly}
}
\value{
A data.frame containing three columns (contig /cov/erage, contig /len/ength and contig /GC/-content). Row names are contig identifiers. Returns results for the common set of contigs between contigSummData and 454LargeContigs.fna.
}
\description{
Extract GC via contig FASTA file and incorporate into len-cov data.frame
}
\examples{
\dontrun{a<-extract.contig.data()}
\dontrun{a1<-add.gc(a,"454LargeContigs.fna")}
}

