% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng.R
\name{extract.all.neighbourhoods}
\alias{extract.all.neighbourhoods}
\title{Extract all k-neighbourhoods from an RNG}
\usage{
extract.all.neighbourhoods(rng,k)
}
\arguments{
\item{rng}{is an \code{igraph} as output from \code{make.rng.from.2d.binning.space}}
\item{k}{a postive integer to select the size of the neighourhood}
}
\value{
A list whose elements contain the contig memberships of each neighbourhood.
}
\description{
This function generates a contig memberships fro each neighbourhood of an RNG
}


