% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-newbler.R
\name{extract.contig.data}
\alias{extract.contig.data}
\title{Extract length and coverage data from first part of 454ContigGraph.txt.}
\usage{
extract.contig.data()
}
\value{
A data.frame containing two columns (contig /cov/erage and contig /len/ength). Row names are contig identifiers.
}
\description{
Extract length and coverage data from first part of 454ContigGraph.txt.
}
\examples{
\dontrun{a<-extract.contig.data()}
}

