% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng.R
\name{prep.for.rng}
\alias{prep.for.rng}
\title{Prepare coverage-GC binning data for calculation of relative neighbourhood graph}
\usage{
prep.for.rng(bDf,minl)
}
\arguments{
\item{bDf}{is a data.frame containing columns named coverage (cov), length (len) and GC (gc)}
\item{minl}{is a integer, being the inclusion threshold for minimum contig length}
}
\value{
A matrix containing two columns: lgcov (log10 coverage from \code{bDf$cov}) and gc (\code{cov$gc}). Rows index contigs, filtered for length via \code{minl}
}
\description{
take a 3-column (cov,len,gc) data.frame (\code{bDf}), subset by contig length (\code{minl}) and return a matrix of log10(cov) and gc data for use in \code{make.rng.from.2d.binning.space}
}


