% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng.R
\name{summarise.neighbourhoods}
\alias{summarise.neighbourhoods}
\title{Summarise data about k-neighbourhoods of an RNG}
\usage{
summarise.neighbourhoods(nhMembershipList,bDf)
}
\arguments{
\item{nhMembershipList}{is a list, output from \code{extract.all.neighbourhoods}}
\item{bDf}{is a matrix output from \code{prep.for.rng}}
}
\value{
A table summarising the size, log10 coverage mean and s.d., GC mean and s.d.
}
\description{
Summary data for neighbourhoods from an RNG made from two binning variables
}


